fix(web): validate a turn body's facts with the protocol schemas before folding them
The console's decoder checked only a discriminator and cast the rest, so a malformed body —
`{"linear":{"issue":null}}`, `{"codehost":{"provider":"github"}}` — passed and threw inside the
formatter when the fold opened. The fact schemas now live in a bundler-safe protocol leaf
(`@agentconnect.md/protocol/user-turn-body`, re-exported beside the tool and plan bodies) and
the decoder validates each fact whole with `safeParse`, dropping the one that fails and keeping
the other.
